tests/contracts/requirements.txt
# Contract Testing Dependencies

# Pact for consumer-driven contract testing
pact-python>=2.0.0

# OpenAPI validation
openapi-spec-validator>=0.7.0
prance>=23.0.0

# API testing
requests>=2.31.0
aiohttp>=3.9.0

# YAML processing
PyYAML>=6.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Contract diff and comparison
deepdiff>=6.7.0
jsonschema>=4.20.0

# Reporting
jinja2>=3.1.0
markdown>=3.5.0
